pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oncapintada"
version = "25.11.6"
authors = [
  { name="Leandro Seixas", email="leandro.seixas@proton.me" },
]
description = "Onça-pintada is a python package designed for the analysis and simulation of atomic configurations in materials science."
readme = "README.md"
license = {file = "LICENSE"} # MIT License
requires-python = ">=3.10"
dependencies = [
    'numpy',
    'pandas',
    'ase',
    'torch',
    'scipy',
    'matplotlib',
    'pytest',
    'pyyaml'
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics"
]

[project.urls]
"Homepage" = "https://github.com/seixasgroup/oncapintada"
